Add page editor with GrapeJS integration
Created a new 'page-editor' directory that includes a very basic website builder. The directory includes the new component 'GrapesJS_Init', containing the necessary code to integrate GrapeJS as the editor for the page. Additionally, some styling changes have been made to enhance the overall look and feel of the page.