Added links to editor
The editor has been updated to show an icon for all blocks that the user can see. On additional block is the Link in which it can be selected and the text updated and the href link as well. In addition when highlighting text user can also make text into link.